requirements.txt
ipykernel
torch
torchvision
visdom
python-telegram-bot
python-dotenv